fix: refactor store for theme and viwer with use createAppStore method support devtools
