+++ src/pages/registration/model.ts
